requirements.txt
aiohttp
python-dotenv
aiosqlite